[PP] denied refs controller spec
