+++ test/deniedrefs/model/DeniedRefsSpec.scala
@@ -16,14 +16,14 @@
 
 package deniedrefs.model
 
+import deniedrefs.TdDeniedRefs
 import play.api.libs.json.Json
 import testsupport.UnitSpec
-import testsupport.testdata.TestData
 
 class DeniedRefsSpec extends UnitSpec {
 
   "json serialization/deserialization" in {
-    val deniedRefs = TestData.deniedRefs1
+    val deniedRefs = TdDeniedRefs.deniedRefs1
 
     val deniedRefsJson = Json.parse(
       //language=JSON
